Revert "Automaticly remove UE4CommandLine.txt when session ended on Android device if UnrealEngine Version >= 4.24"
This reverts commit 2e96e33257a3bf129d22f2ccac1261663850523b.
