+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorLayoutDetails.h
@@ -4,16 +4,28 @@
 
 #include "CoreMinimal.h"
 #include "IDetailCustomization.h"
+#include "Input/Reply.h"
+
+DECLARE_LOG_CATEGORY_EXTERN(LogSpatialGDKEditorLayoutDetails, Log, All);
 
 class FSpatialGDKEditorLayoutDetails : public IDetailCustomization
 {
+private:
+	bool TryConstructMobileCommandLineArgumentsFile(FString& CommandLineArgsFile);
+	bool TryPushCommandLineArgsToDevice(const FString& Executable, const FString& ExeArguments, const FString& CommandLineArgsFile);
+
+	FReply GenerateDevAuthToken();
+	FReply PushCommandLineArgsToIOSDevice();
+	FReply PushCommandLineArgsToAndroidDevice();
+	FReply RemoveCommandLineArgsFromIOSDevice();
+	FReply RemoveCommandLineArgsFromAndroidDevice();
+
+
+	void ForceRefreshLayout();
+
+	IDetailLayoutBuilder* CurrentLayout = nullptr;
+
 public:
 	static TSharedRef<IDetailCustomization> MakeInstance();
 	virtual void CustomizeDetails(IDetailLayoutBuilder& DetailBuilder) override;
-
-private:
-	void ForceRefreshLayout();
-
-private:
-	IDetailLayoutBuilder* CurrentLayout = nullptr;
 };
