+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorModule.h
@@ -1,12 +1,9 @@
 // Copyright (c) Improbable Worlds Ltd, All Rights Reserved
 
-#pragma once
-
 #include "Improbable/SpatialGDKSettingsBridge.h"
 #include "Modules/ModuleManager.h"
 
 class FLBStrategyEditorExtensionManager;
-class FSpatialGDKEditorCommandLineArgsManager;
 
 class FSpatialGDKEditorModule : public ISpatialGDKEditorModule
 {
@@ -42,7 +39,5 @@ private:
 	bool HandleRuntimeSettingsSaved();
 	bool HandleCloudLauncherSettingsSaved();
 
-private:
 	TUniquePtr<FLBStrategyEditorExtensionManager> ExtensionManager;
-	TUniquePtr<FSpatialGDKEditorCommandLineArgsManager> CommandLineArgsManager;
 };
