+++ SpatialGDK/Source/SpatialGDKServices/Public/SpatialCommandUtils.h
@@ -16,5 +16,5 @@ public:
 	SPATIALGDKSERVICES_API static bool StopSpatialService(bool bIsRunningInChina, const FString& DirectoryToRun, FString& OutResult, int32& OutExitCode);
 	SPATIALGDKSERVICES_API static bool BuildWorkerConfig(bool bIsRunningInChina, const FString& DirectoryToRun, FString& OutResult, int32& OutExitCode);
 	SPATIALGDKSERVICES_API static FProcHandle LocalWorkerReplace(const FString& ServicePort, const FString& OldWorker, const FString& NewWorker, bool bIsRunningInChina, uint32* OutProcessID);
-	SPATIALGDKSERVICES_API static bool GenerateDevAuthToken(bool IsRunningInChina, FString& OutTokenSecret, FString& ErrorMessage);
+	SPATIALGDKSERVICES_API static bool GenerateDevAuthToken(bool IsRunningInChina, FString& OutTokenSecret);
 };
